backend/requirements.txt
fastapi==0.103.0
uvicorn[standard]==0.23.2
pydantic==2.5.2
google-generativeai==0.8.3
dashscope==1.20.13
openai==1.54.0
